requirements.txt
requests
shodan
python-nmap
whois
reportlab
flask
sqlmap